third_party/blink/web_tests/virtual/force-defer-script/external/wpt/html/semantics/scripting-1/the-script-element/defer-script/async-script-2-expected.txt

This is a testharness.js-based test.
FAIL Async Script Execution Order Uncaught Error: assert_array_equals: Execution order expected property 0 to be "Inline1" but got "EndOfBody" (expected array ["Inline1", "Sync1", "Async1", "Sync2", "EndOfBody", "DOMContentLoaded"] got ["EndOfBody", "Inline1", "Sync1", "Sync2", "DOMContentLoaded", "Async1"])
Harness: the test ran to completion.
